refactor: move module to `cli/`
