+++ src/actions/disable.ts
@@ -1,8 +1,8 @@
 import chalk from 'chalk';
 import { ESLint } from 'eslint';
 import ora from 'ora';
+import { promptToInputDescription } from '../cli/prompt';
 import { CachedESLint } from '../eslint';
-import { promptToInputDescription } from '../prompt';
 
 export async function doDisableAction(eslint: CachedESLint, results: ESLint.LintResult[], selectedRuleIds: string[]) {
   const description = await promptToInputDescription();
